[ai-assisted] fix(rag): context expansion 리뷰 반영
Issue:

- #290

Why:

PR 리뷰에서 chunk type metadata 변환 방어, 후보 조회 실패 로그, 테스트 중복과 chunkId 상수 중복 정리가 필요하다는 피드백이 있었다.

What:

RagContextBuilder의 chunk type 변환을 fallback 처리하고 WINDOW fallback 의도를 주석으로 남겼다. ChatController는 context expansion 후보 조회 실패 시 warn 로그를 남기고 기존 RAG 결과로 fallback한다. 테스트용 window expander를 공통 fixture로 분리하고 chunkId 상수 중복을 제거했다.

Validation:

- ./gradlew :starter:studio-platform-starter-ai-web:test: PASS

- git diff --check: PASS
